+++ Full-Backend/backend/src/main/java/com/visera/backend/Repository/SkuRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SkuRepository extends JpaRepository<Sku, Integer> {
+public interface SkuRepository extends JpaRepository<Sku, Long> {
 
     Optional<Sku> findBySkuCode(String skuCode);
