+++ Full-Backend/backend/src/main/java/com/visera/backend/Repository/ZoneRepository.java
@@ -2,6 +2,6 @@ package com.visera.backend.Repository;
 import com.visera.backend.Entity.Zone;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ZoneRepository extends JpaRepository<Zone, Integer> {
+public interface ZoneRepository extends JpaRepository<Zone, Long> {
 }
