+++ Full-Backend/backend/src/main/java/com/visera/backend/Service/BinService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface BinService {
     Bin createBin(Bin bin);
-    List<Bin> getBinsByRack(int rackId);
+    List<Bin> getBinsByRack(Long rackId);
 }
